Update dependencies to 0.11
Also, add `purescript` as an explicit dependency.
When people come to work on this project,
it's nice to have a dependency on specific compiler versions.

This gets a little tricky during compiler version changes.
However, it's worlds easier during normal development.

Hopefully, this is an improvement over the current state.